Defer _jsAnalyzer creation
We currently create a `JsAnalyzer` before we ever actually need it. This happens both on boot of VS and when a new project is created. This increases start time and increases memory usage.

I believe that the creation of the analyzer can be safely defered and only created when we actually need it. This change makes this happen. I saw a memory reduction of some tens of MB after the change with the ES6 preview enabled.